Create a hook and component for loading stores/showing loading page
