+++ src/app/collections/Collections.tsx
@@ -1,40 +1,28 @@
-import { DestinyProfileResponse } from 'bungie-api-ts/destiny2';
-import React, { useEffect } from 'react';
+import React from 'react';
 import _ from 'lodash';
-import { DestinyAccount } from '../accounts/destiny-account';
+import { DestinyProfileResponse } from 'bungie-api-ts/destiny2';
 import { D2ManifestDefinitions } from '../destiny2/d2-definitions';
 import './collections.scss';
 import { t } from 'app/i18next-t';
 import ErrorBoundary from '../dim-ui/ErrorBoundary';
-import { D2StoresService } from '../inventory/d2-stores';
 import Catalysts from './Catalysts';
 import { connect } from 'react-redux';
 import { InventoryBuckets } from '../inventory/inventory-buckets';
 import { RootState } from 'app/store/types';
 import { createSelector } from 'reselect';
 import { storesSelector, profileResponseSelector, bucketsSelector } from '../inventory/selectors';
-import { refresh$ } from '../shell/refresh';
 import PresentationNodeRoot from './PresentationNodeRoot';
-import { useSubscription } from 'app/utils/hooks';
 import { useParams } from 'react-router';
 import ShowPageLoading from 'app/dim-ui/ShowPageLoading';
-import { D1StoresService } from 'app/inventory/d1-stores';
-import { queueAction } from 'app/inventory/action-queue';
-import { DimStore } from 'app/inventory/store-types';
-
-interface ProvidedProps {
-  account: DestinyAccount;
-}
 
 interface StoreProps {
-  stores: DimStore[];
   buckets?: InventoryBuckets;
   defs?: D2ManifestDefinitions;
   ownedItemHashes: Set<number>;
   profileResponse?: DestinyProfileResponse;
 }
 
-type Props = ProvidedProps & StoreProps;
+type Props = StoreProps;
 
 function mapStateToProps() {
   const ownedItemHashesSelector = createSelector(storesSelector, (stores) => {
@@ -52,30 +40,15 @@ function mapStateToProps() {
   return (state: RootState): StoreProps => ({
     buckets: bucketsSelector(state),
     defs: state.manifest.d2Manifest,
-    stores: storesSelector(state),
     ownedItemHashes: ownedItemHashesSelector(state),
     profileResponse: profileResponseSelector(state),
   });
 }
 
-function getStoresService(account: DestinyAccount) {
-  return account.destinyVersion === 1 ? D1StoresService : D2StoresService;
-}
-
 /**
  * The collections screen that shows items you can get back from the vault, like emblems and exotics.
  */
-function Collections({ account, stores, buckets, ownedItemHashes, defs, profileResponse }: Props) {
-  useEffect(() => {
-    if (!stores.length) {
-      getStoresService(account).getStoresStream(account);
-    }
-  });
-
-  useSubscription(() =>
-    refresh$.subscribe(() => queueAction(() => getStoresService(account).reloadStores()))
-  );
-
+function Collections({ buckets, ownedItemHashes, defs, profileResponse }: Props) {
   const { presentationNodeHashStr } = useParams();
   const presentationNodeHash = presentationNodeHashStr
     ? parseInt(presentationNodeHashStr, 10)
