+++ src/app/utils/hooks.ts
@@ -1,5 +1,12 @@
 import { useEffect, useState } from 'react';
 import { Subscription } from 'rxjs';
+import { DestinyAccount } from 'app/accounts/destiny-account';
+import { D1StoresService } from 'app/inventory/d1-stores';
+import { D2StoresService } from 'app/inventory/d2-stores';
+import { refresh$ } from 'app/shell/refresh';
+import { queueAction } from 'app/inventory/action-queue';
+import { useSelector } from 'react-redux';
+import { storesLoadedSelector } from 'app/inventory/selectors';
 
 /**
  * Subscribe to an Observable, unsubscribing on changes. Use
@@ -34,4 +41,27 @@ export function useShiftHeld() {
   }, []);
 
   return shiftHeld;
+}
+
+function getStoresService(account: DestinyAccount) {
+  return account.destinyVersion === 1 ? D1StoresService : D2StoresService;
+}
+
+export function useLoadStores(account: DestinyAccount) {
+  const storesLoaded = useSelector(storesLoadedSelector);
+
+  useSubscription(() => {
+    const storesService = getStoresService(account);
+    return refresh$.subscribe(() => queueAction(() => storesService.reloadStores()));
+  });
+
+  useEffect(() => {
+    const storesService = getStoresService(account);
+    if (!storesLoaded) {
+      // TODO: Dispatch an action to load stores instead
+      storesService.getStoresStream(account);
+    }
+  }, [account, storesLoaded]);
+
+  return storesLoaded;
 }
